scripting: fix text capture without active tracer, widen timeout to 64-bit
- Add TrapTracer::captureTextParams() for standalone text capture
  independent of whether tracing is enabled for the trap
- Call captureTextParams() from DoCodeA() when debugger is active
- Widen Token::numValue to uint64_t (fixes timeout truncation)
- Fix traps.def: remove stray 'toolbox' on FillRect param line,
  add missing EqualRect (A8A6) entry

[skip ci]

+++ src/debugger/cmd_parser.cpp
@@ -9,7 +9,7 @@
 #include <cstdio>
 #include <cstring>
 
-bool ParseNumber(std::string_view text, uint32_t &outVal)
+bool ParseNumber(std::string_view text, uint64_t &outVal)
 {
 	if (text.empty()) return false;
 
@@ -18,10 +18,10 @@ bool ParseNumber(std::string_view text, uint32_t &outVal)
 	{
 		auto hex = text.substr(1);
 		if (hex.empty()) return false;
-		uint32_t val = 0;
+		uint64_t val = 0;
 		for (char c : hex)
 		{
-			uint32_t d;
+			uint64_t d;
 			if (c >= '0' && c <= '9')
 				d = c - '0';
 			else if (c >= 'a' && c <= 'f')
@@ -41,10 +41,10 @@ bool ParseNumber(std::string_view text, uint32_t &outVal)
 	{
 		auto hex = text.substr(2);
 		if (hex.empty()) return false;
-		uint32_t val = 0;
+		uint64_t val = 0;
 		for (char c : hex)
 		{
-			uint32_t d;
+			uint64_t d;
 			if (c >= '0' && c <= '9')
 				d = c - '0';
 			else if (c >= 'a' && c <= 'f')
@@ -60,7 +60,7 @@ bool ParseNumber(std::string_view text, uint32_t &outVal)
 	}
 
 	/* decimal */
-	uint32_t val = 0;
+	uint64_t val = 0;
 	for (char c : text)
 	{
 		if (c < '0' || c > '9') return false;
